Remove the old helper methods from transport
Replace with ItemTransport object, which needs to be instantiated. I
think it can be reused between calls to its methods so there's no need
to create new ones all over the place, but that's what I have done so
far. I'm a bit uncertain about module level variables, and not
everything is contained in a class.