WIP: trying to get category_create_tests to pass
The last error from running the tests was that the TErm table didnt
exist in the test database. So I'm running

    manage.py makemigrations

And it fails with

    AttributeError: 'NoneType' object has no attribute 'unique'

Following the [django docs](https://docs.djangoproject.com/en/1.8/topics/db/models/#abstract-related-name)
I put in:

        related_name="%(app_label)s_%(class)s_term"

But still can't even get a helpful error message.