Fix table not showing the rght question types for Items
Interlally it was using add_term which did just what its name said:
it add a term for the item. There was no check that you could only
have one if the thing was a category. So I added a rudimentary support
for that and creted some more test debt. (list copied in below in case
it gets lost)

Then added the apply_term method on Item (see note about how it sucks
for this to be there) which checks the multiplicity of the taxonomy
of the incoming term and acts apporporately. Finally all taxonomies
currently say they are have optional multiplicity (i.e. categories)
whichi s fine for the question type story.

- [ ] Add tests for data_layer.models.Item.apply_Term
- [ ] refactor tests in rest api to use the api
- [ ] Are the updates to the model tags being saved and output okay? Why does that fail?
- [ ] hid/views/add_Categories should only update category if it has actually changed
- [ ] Test add_Categories (in the hid view) with multiple id/term pairs?
- [ ] Add error reporting to the bit of View/Edit screen that adds categories to items.
- [ ] Add tests for adding categories to more than one item at once via browser
- [ ] We probably need a default value in the category selector to show  that there's initailly no value selected
- [ ] Add tests for what happens when add_categories gets called with nonsense items and category names, should fail gracefully
- [ ] Add transport wrappers for listing the terms in a taxonomy, and update View/Editr screen to use these in place of modls directly
- [ ] Add transport wrappers for creating taxnonmies and terms
- [ ] add_Term in api views should retrn json of the serialized version of the new item -- currently empty dict
- [ ] Term.objects.by_taxonomy() should propbably do select related to get the term back as well, seewhere used in api.add_term