Add select_related to by_taxonomy for terms
This makes it more efficeint for `Item.apply_term` to use `term.taxonomy`
because the taxonomy object is selected Along with the term.
I suspect it will be generally useful to do this because we want to
think of the taxonomy details as part of the definitino of each term.

`Term.objects.by_taxonomy()` should propbably do `select_related()` to get the term back as well, see where used in `api.add_term`
https://trello.com/c/fWrvZDau/77-clear-out-technical-debt-from-working-on-taxonomies