Remove reverse lookups for api urls
The API urls don't need to appear in the system urls.  They code that
uses the api, including the transport layer, can function without it.
This commit removes all the calculation of correct urls using reverse
(which depends on the ruls being included in the systme urls) and
replaces with just spaces -- which seems to work just fine.