Skip duplicate records when importing
Ideally we'd have unique IDs from the data source but we can't guarantee that
so next best thing is to put a unique together validator for the body and
timestamp fields. We do this at the REST API level as this is not possible
for a text field. Unfortunately because MySQL truncates microseconds we need
to have a custom field in the serializer to truncate microseconds before
validation. Shouldn't be a problem when we move to 5.6+ all round