Add sub_location to Message
See https://aptivate.kanbantool.com/b/437721-internews-hid-2018#27566989.

At current state, filtering with the new sub_location field is not
related to the location field on Message because it is not clear what
behaviour should take place. For example, filtering only sub locations
of specific locations or filtering on all sub-locations.