Add configuration for DRC and Uganda instances
This is a bit of quick ad hoc approach still - the fixtures will need
manual loading, and some manual changes in the admin backend (after
loading taxonomoy fixture).

But it feels premature to automate this not knowing more about how many
instances will be supported and what approaches for default setups make
most sense (would we move this to multisite etc).